se validan los inputs de contraseña para que sean iguales y se autoincrementa el id por usuario registrado

+++ src/app/models/modeloRegistro.js
@@ -1,27 +1,45 @@
-// Simulador de datos de usuario
-
 let nombreUsuarioRegistro = document.getElementById('usuarioRegistro');
 let contraseñaUsuarioRegistro = document.getElementById('contraseñaRegistro');
 let confirmarContraseñaUsuarioRegistro = document.getElementById('confirmarContraseña');
 
-let botonRegistro = document.getElementById('registroBtn');
+const botonRegistro = document.getElementById('registroBtn');
 
 let listaUsuariosRegistrados = [];
 
+let idCounter = 1;
 botonRegistro.addEventListener('click', () => {
-    let usuario = {
-        id: 1,
-        nombrePersona: nombreUsuarioRegistro.value,
-        contraseña: confirmarContraseñaUsuarioRegistro.value,
-        nombreCuenta: 'cuenta de '+nombreUsuarioRegistro.value,
-        saldo: 200000
-    }
 
-    listaUsuariosRegistrados.push(usuario);
-    localStorage.setItem("storageUsuarioRegistrados", JSON.stringify(listaUsuariosRegistrados));
-    console.log('en este momento te has regsitrado y tus datos se han ido a LocalStorage');
+if(contraseñaUsuarioRegistro.value === confirmarContraseñaUsuarioRegistro.value){
 
-    // limpiar formulario
+            let usuario = {
+                id: idCounter++,
+                nombrePersona: nombreUsuarioRegistro.value,
+                contraseña: confirmarContraseñaUsuarioRegistro.value,
+                nombreCuenta: 'cuenta de '+nombreUsuarioRegistro.value,
+                saldo: 200000
+            }
+    
+            listaUsuariosRegistrados.push(usuario);
+            localStorage.setItem("storageUsuarioRegistrados", JSON.stringify(listaUsuariosRegistrados));
+
+    Swal.fire({
+        position: "center",
+        icon: "success",
+        title: "Felicidades! ya haces parte de ATM-bank",
+        showConfirmButton: false,
+        timer: 2000
+    });
+
+}else{
+
+    Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Las contraseñas No coinciden!",
+        //footer: '<a href="#">Why do I have this issue?</a>'
+    });
+
+}
     nombreUsuarioRegistro.value = '';
     contraseñaUsuarioRegistro.value = '';
     confirmarContraseñaUsuarioRegistro.value = '';
